version 0.1.6
Refactor extension and Python validator (Weslley's feedback)

Extension:
- Split extension.ts into modules: validation.ts, hover.ts, config.ts,
  dictionary.ts, types.ts; extension.ts only registers commands and init.
- Dictionary download delegated to Python script (single implementation).
- Parse script stdout for validation result or script-failure JSON; show
  user-facing messages by error code (DICT_NOT_FOUND, CIF_NOT_FOUND,
  DOWNLOAD_ERROR).

Python:
- Add protocol.py: ValidationResult/ValidationErrorItem, ScriptFailure,
  ErrorCode (1–99), validation_result_from_errors(), script_failure_dict(),
  download_success_dict(). Script prints JSON on success and on failure.
- Add mmcif_types.py: ItemValue, ValidationError, MmCIFValidatorError,
  DictionaryNotFoundError, CifNotFoundError, DownloadError.
- Add download.py: get_cache_dir(), get_cached_dictionary_path(),
  download_dictionary(url, cache_path=). Cache path matches extension
  (tempdir/mmcif-validator-cache/mmcif_pdbx.dic).
- CLI: add subcommand "download-dictionary --url URL [--output PATH]";
  prints {"path": "..."} or script-failure JSON.
- Split validate_mmcif.py: DictionaryParser → dict_parser.py,
  MmCIFParser → cif_parser.py, MmCIFValidator → validator.py.
  validate_mmcif.py imports from these and keeps ValidatorFactory,
  validate(), main(), cmd_download_dictionary().

pyproject.toml: list protocol, mmcif_types, download, dict_parser,
cif_parser, validator in py-modules.
