+++ README.md
@@ -1,6 +1,6 @@
 # PDBe mmCIF Validator
 
-**Version 0.1.5**
+**Version 0.1.6**
 
 Real-time VSCode extension and standalone Python script for validating mmCIF/CIF files against the PDBx/mmCIF dictionary or any CIF dictionary.
