+++ vscode-extension/README.md
@@ -1,6 +1,6 @@
 # PDBe mmCIF Validator
 
-**Version 0.1.5**
+**Version 0.1.6**
 
 A Visual Studio Code extension to validate mmCIF/CIF files against the PDBx/mmCIF dictionary (or any CIF dictionary) with real-time error checking.
