+++ vscode-extension/python-script/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pdbe-mmcif-validator"
-version = "0.1.5"
+version = "0.1.6"
 description = "Validate mmCIF/CIF files against the PDBx/mmCIF dictionary or any CIF dictionary"
 readme = "README.md"
 license = { text = "MIT" }
@@ -43,4 +43,4 @@ Repository = "https://github.com/PDBeurope/mmcif-validator"
 validate-mmcif = "validate_mmcif:main"
 
 [tool.setuptools]
-py-modules = ["validate_mmcif"]
+py-modules = ["validate_mmcif", "protocol", "mmcif_types", "download", "dict_parser", "cif_parser", "validator"]
